+++ menu.php
@@ -1,6 +1,9 @@
 <?php
 include 'koneksi.php';
 
+if (isset($_SESSION['role'])) {
+$role = $_SESSION['role'];
+
 $sql = "SELECT * FROM menu";
 $result = $koneksi->query($sql);
 
@@ -12,13 +15,18 @@ if ($result->rowCount() > 0) {
         echo '<div class="card-body">';
         echo '<div class="row">';
         echo '<div class="col">' . $nomor . '</div>';
-        echo '<div class="col"><img src="' . $row['file_path'] . '" alt="' . $row['nama_makanan'] . '" width="200"></div>';
+
+        echo '<div class="col"><img src="' . $row['file_path'] . '" alt="' . $row['nama_makanan'] . '" width="150"></div>';
         echo '<div class="col"><p>' . $row['nama_makanan'] . '</p></div>';
         echo '<div class="col"><p>' . $row['category'] . '</p></div>';
-        echo '<div class="col"><p>Rp ' . number_format($row['harga'], 0, ',', '.') . '</p></div>';
-        echo '<div class="col"><p>' . $row['deskripsi'] . '</p></div>';
-        echo '<div class="col"><a href="edit_menu.php?id=' . $row['id'] . '" class="btn bg-gradient-danger">Edit</a></div>';
-        echo '<div class="col"><a href="deletemenu.php?id=' . $row['id'] . '" class="btn bg-gradient-info">Delete</a></div>';
+
+        if($role == 'Admin')
+        {
+            echo '<div class="col"><p>Rp ' . number_format($row['harga'], 0, ',', '.') . '</p></div>';
+            echo '<div class="col"><p>' . $row['deskripsi'] . '</p></div>';
+            echo '<div class="col"><a href="edit_menu.php?id=' . $row['id'] . '" class="btn bg-gradient-danger">Edit</a></div>';
+            echo '<div class="col"><a href="deletemenu.php?id=' . $row['id'] . '" class="btn bg-gradient-info">Delete</a></div>';
+        }
         echo '</div>';
         echo '</div>';
         echo '</div>';
@@ -28,5 +36,6 @@ if ($result->rowCount() > 0) {
     }
 } else {
     echo 'Tidak ada data menu yang tersedia.';
+}
 }
 ?>
